requirements.txt
anthropic
langchain
langchain-community
langchain-anthropic
streamlit
yfinance
pandas
numpy
newsapi-python
python-dotenv
matplotlib
plotly
fastapi
uvicorn
pydantic
starlette
python-multipart 